+++ app/src/main/java/yong/tank/modal/tankBascInfo.java
@@ -10,7 +10,7 @@ public class TankBascInfo {
     private int type;
     private int blood;
     private int speed;
-    private int intervalSpeed;
+    private float intervalSpeed;
     private int power;
     private int picture;
     private String tankName; //不联网为坦克名，联网则为自定义名字
@@ -63,8 +63,8 @@ public class TankBascInfo {
     public int getSpeed() {
         return speed;
     }
-    public int getIntervalSpeed() {
-        this.intervalSpeed = (int)((float) StaticVariable.LOCAL_SCREEN_WIDTH /(float)(2*StaticVariable.LOGICAL_FRAME*this.speed/10));
+    public float getIntervalSpeed() {
+        this.intervalSpeed = ((float) StaticVariable.LOCAL_SCREEN_WIDTH /(float)(2*StaticVariable.LOGICAL_FRAME*this.speed/10));
         return this.intervalSpeed;
     }
